backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.31",
    "stripe>=14.2.0",
]
